+++ crates/re_viewer/src/ui/mod.rs
@@ -1,4 +1,3 @@
-mod blueprint_panel;
 mod mobile_warning_ui;
 mod override_ui;
 mod recordings_panel;
@@ -12,7 +11,6 @@ pub(crate) mod query_range_ui;
 pub(crate) mod selection_panel;
 pub(crate) mod space_view_space_origin_ui;
 
-pub use blueprint_panel::blueprint_panel_ui;
 pub use recordings_panel::recordings_panel_ui;
 // ----
